Added Transformer class
