+++ lib/visitor/src/printer.ts
@@ -2,7 +2,7 @@ import { ASTBuilder } from "./sourceBuilder";
 import { SourceKind } from "./ast";
 import { Transformer } from "./transformer";
 
-export class Printer extends Transformer {
+export default class Printer extends Transformer {
   afterParse(): void {
     const files = this.parser.program.sources.filter(
       _source => _source.sourceKind == SourceKind.USER_ENTRY
