+++ lib/visitor/tests/assembly/f32.ts
@@ -3,7 +3,7 @@ const x32: f32 = 1.9;
 const y32: f32 = -1.99;
 
 function foo32(f: f32): f32 {
-  return x + y + f;
+  return x32 + y32 + f;
 }
 
 foo32(2);
