+++ lib/visitor/tests/test.js
@@ -130,20 +130,6 @@ async function loadModule(path) {
 let asc = require("../../../cli/asc");
 
 
-async function testFloatDetection(file){
-  return new Promise( (resolve, reject)=>
-    asc.main([file, "--runtime", "none", "--noEmit"], error => {
-      if (error == null || !/Floating point numbers are not allowed/g.test(error.message)) {
-        if (error == null){
-          error = " Couldn't find float";
-        }
-        reject(file + error);
-        
-      }
-      resolve()
-    }));
-}
-
 (async function() {
     const module = await loadModule('./out/test.wasm');
     await module.runTest();
@@ -163,8 +149,7 @@ async function testFloatDetection(file){
         {"foo":123,"bar":1,"u64Val":"4294967297","i64Val":"-64","flag":false,"baz":"123","uint8array":null,"arr":null,"u32Arr":null,"i32Arr":null, "uint8arrays":null, u64Arr: null, u64_zero:"0"});
     assert.deepEqual(await module.stringOrNull(), null);
     assert.deepStrictEqual(await module.stringAliasTest({str:"Hello"}), "Hello World");
-    await testFloatDetection("assembly/f32.ts");
-    await testFloatDetection("assembly/f64.ts");
+
 })().catch(e => {
     console.error('Error during test execution:', e);
     if (e.code == 'ERR_ASSERTION') {
