+++ lib/visitor/webpack.config.js
@@ -3,7 +3,10 @@ const webpack = require("webpack");
 
 // Build the transformer
 let config = {
-  entry: "./src/index.ts" ,
+  entry: {
+   jsonbindings:"./src/index.ts",
+   ASTPrinter: "./src/printer.ts"
+  },
   module: {
     rules: [
       {
@@ -17,7 +20,7 @@ let config = {
     extensions: [ ".ts", ".js" ]
   },
   output: {
-    filename: "transformerBundle.js",
+    filename: "[name].js",
     path: path.resolve(__dirname, "dist"),
     library: "transformer",
     libraryTarget: "umd",
